+++ MacPacker/Features/About/AboutView.swift
@@ -1,14 +1,15 @@
 //
 //  AboutView.swift
-//  FileFillet
+//  MacPacker
 //
-//  Created by Stephan Arenswald on 03.11.22.
+//  Created by Stephan Arenswald on 11.01.26.
 //
 
 import SwiftUI
 
 struct AboutView: View {
     @Environment(\.openURL) private var openURL
+    @State private var defaultTab = 1
     
     var body: some View {
         VStack (alignment: .center, spacing: 8) {
@@ -21,18 +22,27 @@ struct AboutView: View {
             Text("Version v\(Bundle.main.appVersionLong)", comment: "This text shows the current version of the app in Welcome and About window")
                 .foregroundColor(.secondary)
             
-            Text("\(Bundle.main.appName) has been brought to you by", comment: "Below of this text is a list of people who contributed to the development of this app")
-                .fontWeight(.semibold)
-                .padding(.top, 14)
-            HStack(spacing: 14) {
-                VStack(alignment: .trailing, spacing: 0) {
-                    Text(verbatim: "Stephan Arenswald")
+            if #available(macOS 15.0, *) {
+                TabView(selection: $defaultTab) {
+                    WelcomeOtherProjects()
+                        .tabItem {
+                            Text("More apps", comment: "Title of the tab in the welcome screen that links to other projects.")
+                        }
+                        .tag(1)
                 }
-                VStack(alignment: .leading, spacing: 0) {
-                    Text("idea, code", comment: "This text describes the role of one of the app's developers")
+                .tabViewStyle(.grouped)
+                .padding(.top, 16)
+            } else {
+                TabView(selection: $defaultTab) {
+                    WelcomeOtherProjects()
+                        .tabItem {
+                            Text("More apps", comment: "Title of the tab in the welcome screen that links to other projects.")
+                        }
+                        .tag(1)
                 }
+                .tabViewStyle(.automatic)
+                .padding(.top, 16)
             }
-            .font(.caption2)
             
             #if !STORE
             Text("Support the development...", comment: "Hint to the user to support the app's development via some donation")
@@ -58,36 +68,14 @@ struct AboutView: View {
             Spacer()
             #endif
             
-            Text("Reach out...", comment: "Hint in the About view to let the user contact the developer. The various ways to contact the developer are listed below.")
-                .fontWeight(.semibold)
-                .padding(.top, 14)
-            HStack(spacing: 14) {
-                Link(destination: URL(string: "mailto:\(Constants.supportMail)")!) {
-                    Text(verbatim: "\(Constants.supportMail)")
-                }
-                Link(destination: URL(string: "https://twitter.com/sarensw")!) {
-                    Text(verbatim: "@sarensw")
-                }
-                Link(destination: URL(string: "https://macpacker.app/?ref=about")!) {
-                    Text(verbatim: "macpacker.app")
-                }
+            Button {
+                AckWindowController().show()
+            } label: {
+                Text("Acknowledgements", comment: "Button text to open the window with the list of attributions for other open-source libraries used for the app")
             }
-            
-            HStack(spacing: 0) {
-                Text(verbatim: "\(Calendar.current.component(.year, from: Date())) Stephan Arenswald. ")
-                Text("Published as Open Source under GPL.", comment: "Hint about how the app is published as Open Source using the GPL license")
-            }
-            .font(.footnote)
-            .foregroundColor(.secondary)
-            .padding(.top, 14)
+            .buttonStyle(.plain)
+            .foregroundStyle(.tertiary)
         }
         .padding()
-    }
-}
-
-struct AboutView_Previews: PreviewProvider {
-    static var previews: some View {
-        AboutView()
-            .frame(width: 460, height: 480)
     }
 }